Improve config-gen.sh
First make it so that we only envsubst variables that are provided to us.
This means that if you have a patch with $delete in it it doesn't get
trampled by the envsubst.

Also split up config generation. talosctl will generate both the
controlplane and worker configs internally even if you provide
--output-types. This means that a patch (say a delete) that is only
compatible with one of the configs will fail to apply even though that
final config would never be outputted anyway.

To avoid this we generate the whole config first and then apply all the
patches on top.