Add script to bootstrap flux and 1password secrets
This is needed to turn up the whole cluster. Flux needs it's git keys to
go and configure everything and 1password keys need to be injected so
that onepassword-connect will function and provide all the ExternalSecrets
that are declared in flux that are used by the rest of the cluster.