Giant refactor to split out Kustomizations better.
First changed the talos bootstrap to only target the kubernetes/flux/
directory and it is responsible for configuring the rest of the needed
flux stuff.

Next I refactored how we setup the apps so that they are all wrapped up
in their own individual Flux Kustomize nodes so they fail independently
and we can specify ordering.

The intended pattern currently is:
 apps/<CATEGORY>/
   - kustomization.yaml: this points to subdirectory .ks files
   <DEPLOYMENT>/ deploys the component with the specified name
     - ks.yaml: this is a flux kustomize which installs the `app`
     - app/: this subdir contains the components manifest generation
        - kustomize.yaml the top level kustomize for this app